builds

+++ main/cat_battle_msgs.h
@@ -24,7 +24,7 @@ enum rps_values {
 void send_ping(uint32_t arbid, uint8_t ping_type, uint8_t my_id);
 void send_challenge(uint32_t arbid, uint8_t target_id, uint8_t chal_type);
 uint8_t get_rps();
-void send_rps(uint32_t arbid, uint8_t target_id, uint8_t rps_type, uint8_t value);
-bool didWinRPS(uint8_t my_rps, uint8_t tar_rps);
+void send_rps(uint32_t arbid, uint8_t target_id, uint8_t rps_type, int16_t value);
+bool didWinRPS(int16_t my_rps, int16_t tar_rps);
 
 #endif /* CATBATTLEMSGS_H_ */
